pyproject.toml
[tool.poetry]
name = "tech-radar"
version = "1.0.0"
description = "The simplest tech radar creator in python."
authors = ["Daniel Gutson <0x123456789@gmail.com>"]
license = "GPL-3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.2"
matplotlib = "^3.9.2"
numpy = "^2.1.0"

[tool.poetry.scripts]
tech-radar = "tech_radar.radar:main"

[tool.poetry.group.dev.dependencies]
mypy = "^1.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
